# contracts/http-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: VÖLUND catalog API
  version: 1.0.0
  description: Versioned CAD metadata access with explicit managed source moves.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/sessions:
    get:
      operationId: listOwnSessions
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Active sessions owned by the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnSession"
        "401":
          description: No active session exists.
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session metadata; opaque credentials are returned as cookies only.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Credentials are invalid.
...
  /api/v1/sessions/{sessionId}:
    delete:
      operationId: revokeOwnSession
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The selected non-current session was revoked.
        "401":
          description: No active session exists.
        "403":
          description: CSRF verification failed.
        "404":
          description: The session is not owned, is current, or is no longer active.
  /api/v1/session:
    get:
      operationId: getCurrentSession
      responses:
        "200":
          description: The current authenticated browser session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentSession"
        "401":
          description: No active session exists.
    delete:
      operationId: logout
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Session revoked and cookies cleared.
        "401":
          description: No active session exists.
        "403":
          description: CSRF verification failed.
  /api/v1/account/password:
    put:
      operationId: changeOwnPassword
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeOwnPasswordRequest"
      responses:
        "204":
          description: Password changed and other sessions revoked.
        "400":
          description: The current or new password is invalid.
  /api/v1/users:
    get:
      operationId: listUsers
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Local user inventory for owners and administrators.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedUser"
        "403":
          description: User-administration capability is required.
    post:
      operationId: createUser
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: The active local account was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          description: Account fields are invalid.
        "403":
          description: User-administration or owner capability is required.
        "409":
          description: The email address is already registered.
  /api/v1/users/{userId}:
    patch:
      operationId: updateUser
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated local account.
  /api/v1/users/invitations:
    post:
      operationId: inviteUser
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserRequest"
      responses:
        "200":
          description: Invited account plus its one-time activation secret.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCreated"
        "403":
          description: User-administration or owner capability is required.
        "409":
          description: The email address is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "403":
          description: The requested role or status change is forbidden.
        "404":
          description: User not found.
        "409":
          description: The last active owner cannot be disabled or demoted.
  /api/v1/users/{userId}/password:
    post:
      operationId: resetUserPassword
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetUserPasswordRequest"
      responses:
        "204":
          description: Credential replaced and stale sessions revoked.
        "400":
          description: Password is invalid.
        "403":
          description: The target account cannot be administered by this actor.
        "404":
          description: User not found.
  /api/v1/settings:
    get:
      operationId: listSettings
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Complete code-owned setting registry with effective values.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"
...
  /api/v1/settings/{settingKey}:
    put:
      operationId: updateSetting
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: settingKey
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingRequest"
      responses:
        "200":
          description: Validated persisted setting.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
        "400":
          description: The value is invalid for this setting.
        "403":
          description: Settings-administration capability is required.
        "404":
          description: The key is not registered.
        "409":
          description: The submitted revision is stale.
...
  /api/v1/libraries:
    get:
      operationId: listManagedLibraries
      responses:
        "200":
          description: Administrative library details including canonical host paths.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedLibrary"
        "403":
          description: Library-administration capability is required.
    post:
      operationId: createManagedLibrary
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLibraryRequest"
      responses:
        "200":
          description: Registered library with an immutable key and canonical path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedLibrary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Library-administration capability is required.
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/libraries/validate:
    post:
      operationId: validateLibraryPath
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateLibraryPathRequest"
      responses:
        "200":
          description: Canonical path and permission observations from the daemon host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryPathValidation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Library-administration capability is required.
  /api/v1/libraries/{rootKey}:
    patch:
      operationId: updateManagedLibrary
      description: Requires the revision loaded before confirmation. Stale edits must reload and be confirmed again.
      parameters:
        - $ref: "#/components/parameters/RootKey"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLibraryRequest"
      responses:
        "200":
          description: Updated safe metadata and activation state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedLibrary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Library-administration capability is required.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The library revision changed; reload before retrying.
        "422":
          description: Missing or incorrectly typed expectedRevision.
  /api/v1/libraries/{rootKey}/scans:
    post:
      operationId: startLibraryScan
      parameters:
        - $ref: "#/components/parameters/RootKey"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartLibraryScanRequest"
      responses:
        "202":
          description: Durable scan job accepted or coalesced with active work for this library.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryScanResult"
        "403":
          description: Library-administration capability is required.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/quarantines:
    get:
      operationId: listSourceQuarantines
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Bounded administrative inventory without host paths.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuarantinePage"
        "403":
          description: Administrative capability is required.
...
  /api/v1/lifecycle/preview:
    post:
      operationId: previewCatalogLifecycle
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecyclePreviewRequest"
      responses:
        "200":
          description: Short-lived actor, resource, and revision-bound impact plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecyclePlan"
        "403":
          description: The role cannot preview this lifecycle action.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/models/{modelId}/files:
    get:
      operationId: listModelFiles
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Bounded stable fileset page ordered by primary state, ordinal, and path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelFilePage"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: volund_session
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
          maxLength: 1024
...
    CurrentSession:
      type: object
      required: [sessionId, userId, email, displayName, role, mustChangePassword]
      properties:
        sessionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [owner, administrator, editor, viewer]
        mustChangePassword:
          type: boolean
    OwnSession:
      type: object
      required:
        - id
        - current
        - createdAtUnixMs
        - lastSeenAtUnixMs
        - idleExpiresAtUnixMs
        - absoluteExpiresAtUnixMs
        - clientAddress
        - userAgent
      properties:
        id:
          type: string
          format: uuid
        current:
          type: boolean
        createdAtUnixMs:
          type: integer
        lastSeenAtUnixMs:
          type: integer
        idleExpiresAtUnixMs:
          type: integer
        absoluteExpiresAtUnixMs:
          type: integer
        clientAddress:
          type: [string, "null"]
        userAgent:
          type: [string, "null"]
...